Add o arquivo conftest.
close #19
