+++ libpythonpro/tests/test_spam/test_usuarios.py
@@ -1,22 +1,4 @@
-import pytest
-
-from libpythonpro.spam.db import Conexao
 from libpythonpro.spam.modelos import Usuario
-
-
-@pytest.fixture(scope='module')
-def conexao():
-    conexao_obj = Conexao()
-    yield conexao_obj
-    conexao_obj.fechar()
-
-
-@pytest.fixture()
-def sessao(conexao):
-    sessao_obj = conexao.gera_sessao()
-    yield sessao_obj
-    sessao_obj.roll_back()
-    sessao_obj.fechar()
 
 
 def test_salvar_usuario(sessao):
